Makes create_human_player_binding return a dict of boolean values
So now the action that was being used to move the crosshair to a
specific position was changed to the 'ch_mouse' action, which either
moves or not the crosshair to the exact mouse position. THis is useful
so that the dictionary returned by the decide_action method for every
player can simply be made of boolean values.