Changes architecture of Game. Players don't have decide_action_fun anymore
Now the user of the Game class is responsible for computing the actions
for each player before calling the update_physics function of the Game
class. This is necessary for general RL agents to interface with this
game in an easy way.

The code for the random player, simple AI and not so simple AI was
updated to reflect these changes, and so did the example.py file